update bt client/server
