--- communication/KEYBOARD_CONFIG.py
@@ -1,16 +1,0 @@
-# keyboard library key mappings
-# current defined for using the arrow keys (201911091541)
-FORWARD = "up"
-BACKWARD = "down"
-LEFT = "left"
-RIGHT = "right"
-
-KEYBOARD_CONFIG_DICT = {
-    'FORWARD': FORWARD,
-    'BACKWARD': BACKWARD,
-    'LEFT': LEFT,
-    'RIGHT': RIGHT
-}
-
-# how often should commands be sent from the controller
-DEBOUNCE_INTERVAL = 0.2 # seconds
