--- communication/SWARMER_BT_INFO.py
@@ -1,14 +1,0 @@
-SWARMER_ID_DICT = {
-    "S1": 'B8:27:EB:19:6A:75',
-    "S2": 'B8:27:EB:6A:D3:2D',
-    "S3": 'B8:27:EB:0D:A8:B9'
-}
-SWARMER_ADDR_DICT = {
-    'B8:27:EB:19:6A:75': 'S1',
-    'B8:27:EB:6A:D3:2D': 'S2',
-    'B8:27:EB:0D:A8:B9': 'S3'
-}
-
-UUID = "94f39d29-7d6d-437d-973b-fba39e49d4ee"
-
-ALL_ADDRESSES = list(SWARMER_ADDR_DICT.keys())
